Move logic to container and fix height
